+++ CHANGELOG.md
@@ -5,5 +5,12 @@
 - Implemented versioning footer.
 - Set up basic Next.js template
 
+## [0.2.0] - 2025-02-25
+- updated metadata and favicon
+- added relevant  content
+- setup github action versioning workflow
+- store zustand state managent
+- added theme provider and toggle theme button
+